+++ NEWS.md
@@ -1,4 +1,4 @@
-# targets 1.3.2.9002 (development)
+# targets 1.3.2.9003 (development)
 
 ## Invalidating changes